Add read timeouts to connections; Set max concurrent handshakes to less than maximum mio channel capacity; Switch level triggers to edge triggers to fix issue with partial reads; Add tests
